+++ src/img/test/basics.test.js
@@ -40,6 +40,18 @@ var before = tap4nodeunit.before;
 var test = tap4nodeunit.test;
 
 
+function objCopy(obj, target) {
+    if (!target) {
+        target = {};
+    }
+    Object.keys(obj).forEach(function (k) {
+        target[k] = obj[k];
+    });
+    return target;
+}
+
+
+
 test('imgadm --version', function (t) {
     exec('imgadm --version', function (err, stdout, stderr) {
         t.ifError(err, err);
@@ -65,8 +77,8 @@ test('imgadm --version', function (t) {
 });
 
 
-test('imgadm -vv   # bunyan debug log on stderr', function (t) {
-    exec('imgadm -vv bogus', function (err, stdout, stderr) {
+test('imgadm -v   # bunyan debug log on stderr', function (t) {
+    exec('imgadm -v bogus', function (err, stdout, stderr) {
         t.ok(err);
         t.equal(err.code, 1);
         t.equal(stdout, '', 'stdout');
@@ -79,8 +91,11 @@ test('imgadm -vv   # bunyan debug log on stderr', function (t) {
     });
 });
 
-test('imgadm -vvv   # bunyan "src" log on stderr', function (t) {
-    exec('imgadm -vvv bogus', function (err, stdout, stderr) {
+test('IMGADM_LOG_LEVEL=trace imgadm   # bunyan "src" log on stderr', function (t) {
+    var env = objCopy(process.env);
+    env.IMGADM_LOG_LEVEL = 'trace';
+    var execOpts = {env: env};
+    exec('imgadm bogus', execOpts, function (err, stdout, stderr) {
         t.ok(err);
         t.equal(err.code, 1);
         t.equal(stdout, '', 'stdout');